Make Condition not assignable to ConditionAttributes
The previous breaking change of Condition.and/or could cause hidden errors, where a Condition was mistakenly passed as a ConditionSet,
due to the {[string]: any} extension of ConditionAttributes.
This means that code not adjusting to the breaking change would still compile but fail in runtime.

Changing ConditionAttributes so that Condition.and(...) is not assignable to type ConditionSet, giving a compile error if Condition.and() is
not migrated to ConditionSet.and() as per the breaking change.